+++ CHANGES.md
@@ -5,6 +5,7 @@
 * New / save / close workspace toasts sometimes not displayed [#619](https://github.com/CCI-Tools/cate/issues/619)
 * Missing notification on auto-update [#615](https://github.com/CCI-Tools/cate/issues/615)
 * Show a data source's terms and conditions [#518](https://github.com/CCI-Tools/cate/issues/518)
+* Implement a dimensions' selector dialog [#503](https://github.com/CCI-Tools/cate/issues/503)
 
 
 ## Changes in 2.0.0-dev.9